requirements.txt
# Core
pandas==2.0.3
sqlalchemy==2.0.25
python-dotenv==1.0.1

# Database Driver
psycopg2-binary==2.9.9

# Forecasting
prophet==1.1.5
cmdstanpy==1.2.4
pystan==2.19.1.1

# Airflow providers
apache-airflow-providers-postgres==5.8.0

# Testing
pytest==7.4.4

# Optional but useful
matplotlib==3.5.3   # ✅ works with Python 3.8
plotly
